+++ admin/Changelog/Configuration.hs
@@ -1,0 +1,16 @@
+module Changelog.Configuration where
+
+changesFP :: FilePath
+changesFP = "changes"
+
+bugfixesFP :: FilePath
+bugfixesFP = "bugfixes"
+
+highlightsFP :: FilePath
+highlightsFP = "highlights"
+
+newFP :: FilePath
+newFP = "new"
+
+deprecatedFP :: FilePath
+deprecatedFP = "deprecated"
